+++ src/vs/workbench/contrib/codeEditor/browser/accessibility/accessibility.ts
@@ -9,7 +9,7 @@ import { ConfigurationTarget, IConfigurationService } from 'vs/platform/configur
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { IAccessibilityService } from 'vs/platform/accessibility/common/accessibility';
 import { Action2, registerAction2 } from 'vs/platform/actions/common/actions';
-import { accessibilityHelpIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityContribution';
+import { accessibilityHelpIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibility.contribution';
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
 import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { alert } from 'vs/base/browser/ui/aria/aria';
